requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
google-generativeai==0.3.1
PyMuPDF==1.23.5
python-docx==1.0.1
bson==0.5.10 